Containerized headless cypress tests
