+++ templates/e2e/cypress-sample.json
